pyproject.toml
[project]
name = "pre-org-air-platform"
version = "0.1.0"
description = "AI Readiness assessment for private equity"
authors = [
    {name = "Anirudh Raj, Minal Naranje, Jahanvi Patil"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "snowflake-connector-python (>=4.2.0,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "boto3 (>=1.42.36,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "poetry-plugin-shell (>=1.0.1,<2.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "snowflake-sqlalchemy (>=1.8.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "sec-edgar-downloader (==5.0.2)",
    "pyrate-limiter (==3.1.0)",
    "jobspy (>=0.31.0,<0.32.0)",
    "pandas (==2.2.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "python-jobspy (>=1.1.82,<2.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.14.14,<0.15.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
